requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.1
Werkzeug==3.1.2
python-dotenv==1.0.0
PyPDF2==3.0.0
PyMuPDF
pillow~=11.1.0
pypdf~=5.3.1